myVOD/pyproject.toml
[project]
name = "myvod"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.5.3",
    "dj-database-url>=3.0.1",
    "django>=5.2.7",
    "django-celery-beat>=2.8.1",
    "django-cors-headers>=4.9.0",
    "django-extensions>=4.1",
    "django-redis>=6.0.0",
    "django-stubs>=5.2.6",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.28.0",
    "google-generativeai>=0.8.5",
    "gunicorn>=23.0.0",
    "httpx>=0.28.1",
    "ipython>=9.6.0",
    "pandas>=2.3.3",
    "psycopg2-binary>=2.9.10",
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "pytest-django>=4.11.1",
    "python-dotenv>=1.1.1",
    "redis>=6.4.0",
    "requests>=2.32.5",
    "ruff>=0.13.3",
    "supabase>=2.21.1",
    "ty>=0.0.1a21",
]
